requirements.txt
requests
python-dotenv
tkcalendar
email-validator